Products page

+++ ShopOnline.Models/Dtos/CartItemDto.cs
@@ -8,8 +8,8 @@
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
         public string ProductImageURL { get; set; }
-        public double Price { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal TotalPrice { get; set; }
         public int Qty { get; set; }
     }
 }
